backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
sqlalchemy
boto3
python-multipart
pymupdf4llm
PyMuPDF
chromadb
python-docx
python-pptx
tqdm
regex
psycopg[binary]
qdrant_client
weasyprint
markdown
pandas
openpyxl
redis
psycopg2